moved misc stuff out of database class
